Rename class to Application.java
